Store OrganismType on OrganismBundle
